feat(dedicated.netapp): refactor efs network vrack order link (#11003)
ref: MANAGER-12423
