+++ packages/manager/modules/netapp/src/dashboard/vrack-configuration/service.js
@@ -1,12 +1,9 @@
-import { VRACK_ORDER_URLS } from './constants';
-
 export default class NetAppVrackConfigurationService {
   /* @ngInject */
   constructor($http, $q, Apiv2Service) {
     this.$http = $http;
     this.$q = $q;
     this.Apiv2Service = Apiv2Service;
-    this.VRACK_ORDER_URLS = VRACK_ORDER_URLS;
   }
 
   filterAllowedVrack(vracks, vrackServicesId) {
@@ -32,10 +29,6 @@ export default class NetAppVrackConfigurationService {
     return this.$http
       .get(`/vrack/${vrackId}/allowedServices?serviceFamily=vrackServices`)
       .then(({ data }) => data);
-  }
-
-  getVrackOrderUrl(subsidiary) {
-    return this.VRACK_ORDER_URLS[subsidiary] || this.VRACK_ORDER_URLS.DEFAULT;
   }
 
   linkVrackToVrackServices(vrackId, vrackServicesId) {
